converts parameters to daily resolution

NOTE this currently only works with hard coded variables.

cwatm_hourly_to_daily_ncdf4(inpath, outpath, verbose = FALSE)

Arguments

inpath

filepath to folder containing .nc input files

outpath

filepath to write CWatM input files

verbose

print status?

Value

status code

Details

For CWatM climate variables required in netcdf format:

precipitation (Kg m-2 s-1), variable name = pr_nor2

temperature: max, min & average (K), variable name = tas_nor2, tasmax_nor2, tasmin_nor2

humidity (relative %), variable name = hurs_nor

surface pressure (Pa), variable name = ps_nor

radiation (short wave & long wave downwards) (W m-2), variable name = rsds_nor, rlds_nor,

windspeed (m/s), variable name = wind

You need to pass the following hourly MetNo Variables to this function:

"precipitation_amount" (mm) "air_temperature_2m" (K) "relative_humidity_2m" (%) "air_pressure_at_sea_level" (pa) "integral_of_surface_downwelling_shortwave_flux_in_air_wrt_time" (W/m^2 s)* "integral_of_surface_downwelling_longwave_flux_in_air_wrt_time" (W/m^2 s)** "wind_speed_10m"

(*) "integral_of_Y_wrt_X" means int Y dX. The data variable should have an axis for X specifying the limits of the integral as bounds. "wrt" means with respect to. The surface called "surface" means the lower boundary of the atmosphere. "shortwave" means shortwave radiation. Downwelling radiation is radiation from above. It does not mean "net downward". Surface downwelling shortwave is the sum of direct and diffuse solar radiation incident on the surface, and is sometimes called "global radiation". When thought of as being incident on a surface, a radiative flux is sometimes called "irradiance". In addition, it is identical with the quantity measured by a cosine-collector light-meter and sometimes called "vector irradiance". In accordance with common usage in geophysical disciplines, "flux" implies per unit area, called "flux density" in physics.

(**) This variable has missing data for most of 2014, 2016, and 14 hours in 2019 (see Known data issues)

source: https://github.com/metno/NWPdocs/wiki/MET-Nordic-dataset
Date Accessed: 06-10-2024