Download SeNorge2018 data

This function downloads the files as queried by metnordic_buildquery(). If you do not pass a directory, files will be downloaded in the working directory. You have the option of downloading the following SeNorge2018 variables at daily resolution:

tn (minimum daily temperature in Celsius)

tx (maximum daily temperature in Celsius)

rr (sum daily precipitation in mm),

tg (mean daily temperature in Celsius) If you do not provide a vector of variables to download, all will be downloaded. You can also provide a polygon (Georeferenced simple feature, the same that you provided to metnordic_coordwindow()) to verify that the download is covering your desired area.

Usage

senorge_download(
  queries,
  directory = NULL,
  variables = NULL,
  polygon = NULL,
  verbose = FALSE
)

Arguments

queries

(list) generated by metnordic_buildqueries()

directory

(char) folder where to download (will be created if it does not exist)

variables

(char, vector) variables to download ("tn", "tx", "rr", "tg")

polygon

(char) path to same polygon shapefile as passed to metnordic_coordwindow()

verbose

(flag) print status?

Value

path to downloaded files